Add CDI Extension tests
