+++ tck/pom.xml
@@ -84,6 +84,13 @@
       <scope>provided</scope>
     </dependency>
     
+    <dependency>
+      <groupId>jakarta.enterprise</groupId>
+      <artifactId>jakarta.enterprise.cdi-api</artifactId>
+      <version>${jakarta.enterprise.cdi.version}</version>
+      <scope>provided</scope>
+    </dependency>
+    
     <!-- Provided Entity APIs -->
     <dependency>
       <groupId>jakarta.nosql</groupId>
