+++ tck/src/main/java/ee/jakarta/tck/data/core/cdi/provider/PersonEntity.java
@@ -13,21 +13,18 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
-package ee.jakarta.tck.data.full.cdi;
+package ee.jakarta.tck.data.core.cdi.provider;
 
-import java.util.List;
-
-import jakarta.data.repository.DataRepository;
-import jakarta.data.repository.Repository;
-
-import ee.jakarta.tck.data.full.cdi.provider.PersonExtension;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
 
 /**
- * A Directory repository for testing.
- * 
- * @see ee.jakarta.tck.data.full.cdi.provider.DirectoryRepository
+ * A custom entity annotation for testing.
  */
-@Repository(provider = PersonExtension.PERSON_PROVIDER)
-public interface Directory extends DataRepository<Person, Long> {    
-    List<String> findFirstNameByIdInOrderByAgeDesc(List<Long> ids);
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface PersonEntity {
+
 }
