+++ tck/src/main/java/ee/jakarta/tck/data/core/cdi/provider/PersonExtension.java
@@ -13,7 +13,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
-package ee.jakarta.tck.data.full.cdi.provider;
+package ee.jakarta.tck.data.core.cdi.provider;
 
 import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
