+++ tck/src/main/resources/ee/jakarta/tck/data/framework/signature/sig-test-pkg-list.txt
@@ -24,4 +24,6 @@
 
 jakarta.data
 jakarta.data.exceptions
+jakarta.data.model
+jakarta.data.page
 jakarta.data.repository
